Some improvements to SILE typesetting.
This is mostly a toy floating environment that does some *very* simple
fallback page layout.  If a page is overfull and ends with a floating
thing, or underfull and the next thing is a floating thing, temporarily
remove that from the queue and see if that gives a better layout, then
put it back in the queue at the top of the next page.