Fix toyfloat with \eject.
This is a bit of a hack, but it appears to work:

If we have an under-full page but the last thing in it is a big penalty,
assume that nothing we do will come up with a better solution and early
eject the page.  This fixes having page numbers on blank pages
(especially the title page).